refactor(ipc): split ipc/mod.rs into state + builder + pipeline peers (#597)
Item 6 of the architecture-cleanup roadmap. Pure refactor — no
behaviour change. The largest god-file in the project (2247 lines)
now reads as a 67-line front door over four sibling files.

Before:
- ipc/mod.rs at 2247 lines holding AppState (~250 LOC), AppStateBuilder
  (~390 LOC), build_default + setting parsers + diarizer helper
  (~480 LOC), redirect policy + transcriber loaders + run_pipeline
  (~440 LOC), and ~690 lines of tests.

After:
- ipc/mod.rs: 67 lines — module declarations + public re-exports +
  pub(crate) re-exports for the autostart-mode encode/decode used by
  external callers.
- ipc/state.rs: 833 lines — types (ForegroundApp, DataServices,
  AppState, RuntimeFlags, TranscribeSlot), the production constructor
  AppState::build_default, the swap_transcriber hot-swap, encode/decode
  autostart helpers, parse_*_setting helpers, build_diarizer_inner.
- ipc/builder.rs: 433 lines — AppStateBuilder + impl + the explicit
  build() function used by tests.
- ipc/pipeline.rs: 318 lines — model-download redirect policy
  (RedirectDecision, redirect_decision, is_huggingface_host,
  MAX_DOWNLOAD_REDIRECTS), transcriber loaders
  (load_transcriber_for_model + the pub(super) build_transcriber),
  and the pure run_pipeline orchestrator.
- ipc/tests.rs: 709 lines — every test as a peer module, accessing
  the split modules via super::pipeline::* and super::state::*.

Public API unchanged. lib.rs and ipc::commands::settings continue to
resolve crate::ipc::decode_autostart_mode / encode_autostart_mode via
the pub(crate) re-export in mod.rs. Crate consumers that import
crate::ipc::AppState / AppStateBuilder / run_pipeline /
load_transcriber_for_model continue to work — those names remain at
the crate::ipc:: path via pub use in mod.rs.

Verified locally:
- cargo check --all-targets clean
- cargo clippy --all-targets -- -D warnings clean
- cargo clippy --lib --no-default-features -- -D warnings clean
  (the cross-platform path from CLAUDE.md)
- cargo fmt --all (applied)
- cargo test --lib: 412 pass, 0 fail (same count as before)
